move ab tests into module; configure coverage;
