+++ setup.cfg
@@ -25,8 +25,7 @@ exclude = .git,
 
 [mypy]
 ignore_missing_imports = True
-exclude = 
-    solvis_graphql_api/client
+exclude = solvis_graphql_api/client
 
 [coverage:run]
 # uncomment the following to omit files during running
@@ -34,6 +33,10 @@ omit =
     solvis_graphql_api/pynamodb_config.py
     solvis_graphql_api/handler.py
     solvis_graphql_api/tests/*
+    solvis_graphql_api/client    
+    solvis_graphql_api/scripts/ab_test.py
+    solvis_graphql_api/scripts/cli_ab_test.py
+
 [coverage:report]
 exclude_lines =
     pragma: no cover
